Clean code by Tuan
