--- src/main/java/view/AccManagePanel.java
@@ -1,6 +1,0 @@
-package view;
-
-import javax.swing.*;
-
-public class AccManagePanel extends JPanel {
-}
