--- src/main/java/view/CustomerPanel.java
@@ -1,6 +1,0 @@
-package view;
-
-import javax.swing.*;
-
-public class CustomerPanel extends JPanel {
-}
