--- src/main/java/view/ExportPanel.java
@@ -1,6 +1,0 @@
-package view;
-
-import javax.swing.*;
-
-public class ExportPanel extends JPanel {
-}
