--- src/main/java/view/ImportPanel.java
@@ -1,6 +1,0 @@
-package view;
-
-import javax.swing.*;
-
-public class ImportPanel extends JPanel {
-}
