--- src/main/java/view/NotificationPanel.java
@@ -1,6 +1,0 @@
-package view;
-
-import javax.swing.*;
-
-public class NotificationPanel extends JPanel {
-}
