+++ src/main/java/view/Style.java
@@ -5,7 +5,7 @@ import java.awt.*;
 public class Style {
     public static Color BACKGROUND_COLOR = new Color(0, 128, 255, 150);
     
-    public static Color BACKGROUND_COLOR_DIV2 = new Color(0, 128, 255, 0);
+    public static Color BACKGROUND_COLOR_FUll_BLUR = new Color(0, 128, 255, 0);
     
     public static Font FONT_SIZE = new Font("Arial", Font.PLAIN, 15);
 }
